test(persistence-sqlite): co-locate sqlite_store tests with their submodules (KAN-762)
