+++ crates/kanban-persistence-sqlite/src/sqlite_store/tests/mod.rs
@@ -1,0 +1,12 @@
+mod entities;
+mod graph;
+mod init;
+mod metadata;
+mod persistence_store;
+
+pub(crate) fn make_rt() -> tokio::runtime::Runtime {
+    tokio::runtime::Builder::new_multi_thread()
+        .enable_all()
+        .build()
+        .unwrap()
+}
